refactor(frontend): extract ic-domains generation
