+++ package.json
@@ -14,9 +14,10 @@
 		"build:compress": "./scripts/build.compress.sh",
 		"build:csp": "node scripts/build.csp.mjs",
 		"build:metadata": "node scripts/build.metadata.mjs",
+		"build:ic-domains": "node scripts/build.ic-domains.mjs",
 		"build:ii-alternative-origins": "node scripts/build.ii-alternative-origins.mjs",
 		"build:copy-workers": "node ./scripts/build.copy-workers.mjs",
-		"build:post-process": "npm run build:metadata && npm run build:ii-alternative-origins && npm run build:csp && npm run build:copy-workers && npm run build:compress",
+		"build:post-process": "npm run build:metadata && npm run build build:ic-domains && npm run build:ii-alternative-origins && npm run build:csp && npm run build:copy-workers && npm run build:compress",
 		"build:tokens-sns": "node scripts/build.tokens.sns.mjs && npm run format",
 		"build:tokens-ckerc20": "node scripts/build.tokens.ckerc20.mjs && npm run format",
 		"dev": "vite dev",
