+++ scripts/build.metadata.mjs
@@ -3,28 +3,9 @@
 import { config } from 'dotenv';
 import { readFileSync, writeFileSync } from 'node:fs';
 import { join } from 'node:path';
-import { ENV, findHtmlFiles } from './build.utils.mjs';
+import { ENV, findHtmlFiles, replaceEnv } from './build.utils.mjs';
 
 config({ path: `.env.${ENV}` });
-
-const replaceEnv = ({ content, pattern, value }) => {
-	const regex = new RegExp(pattern, 'g');
-	return content.replace(regex, value);
-};
-
-const parseDomain = (targetFile) => {
-	let content = readFileSync(targetFile, 'utf8');
-
-	// For information such as custom domains, only the domain without protocol is required
-	const { host: value } = new URL(process.env['VITE_OISY_URL'] ?? 'https://oisy.com');
-	content = replaceEnv({
-		content,
-		pattern: `{{VITE_OISY_DOMAIN}}`,
-		value
-	});
-
-	writeFileSync(targetFile, content);
-};
 
 const parseMetadata = (targetFile) => {
 	let content = readFileSync(targetFile, 'utf8');
@@ -93,7 +74,6 @@ htmlFiles.forEach((htmlFile) => parseMetadata(htmlFile));
 
 parseUrl(join(process.cwd(), 'build', 'sitemap.xml'));
 parseMetadata(join(process.cwd(), 'build', 'manifest.webmanifest'));
-parseDomain(join(process.cwd(), 'build', '.well-known', 'ic-domains'));
 
 // SEO
 htmlFiles.forEach((htmlFile) => removeMetaRobots(htmlFile));
